api_service/requirements.txt
Django==4.0.6
psycopg2>=2.8
djangorestframework==3.13.1
requests==2.28.1 